refactor: updated existing components
