+++ src/scripts/components/Navigation/components/Logo.js
