+++ src/scripts/components/Navigation/index.js
@@ -1,0 +1,1 @@
+export { default as Navigation } from './Navigation'
