+++ src/scripts/components/Navigation/svg/DarthVaderSVG.js
