+++ src/scripts/components/Navigation/svg/LogoSVG.js
