+++ src/scripts/pages/Home/Home.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import Navigation from 'components/navigation/Navigation'
+import { Navigation } from 'components/Navigation'
 
 const Home = () => (
     <div>
